+++ CHANGELOG.md
@@ -1226,3 +1226,4 @@
 - [2026-05-17 14:19] refactor(navigation): optimize data flow between domain and data layers
 - [2026-05-17 14:46] feat(core): add bi-directional event bus handler for core
 - [2026-05-17 15:14] fix(security): fix memory leak by cancelling security subscription
+- [2026-05-17 16:06] docs(i18n): update readme and API contracts for i18n